+++ src/frontend/src/common/constants.js
@@ -1,1 +1,6 @@
 export const MAX_INGREDIENT_COUNT = 3;
+
+/* eslint-disable */
+export const EMAIL_REGEX =
+  /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+export const NUMBER_REGEX = /\d/;
